# api-doc/swagger.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/accounts":{"post":{"operationId":"AccountsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountDto"}}}},"responses":{"201":{"description":"返回创建的账户信息","content":{"application/json":{"schema":{"example":{"username":"admin","nickname":"管理员","telephone":"13800138000"}}}}}},"summary":"创建账户","tags":["系统账户管理"]},"get":{"operationId":"AccountsController_findAll","parameters":[],"responses":{"200":{"description":"返回账户列表"}},"summary":"获取账户列表","tags":["系统账户管理"]}},"/api/accounts/currentUser":{"get":{"operationId":"AccountsController_getCurrentUser","parameters":[],"responses":{"200":{"description":""}},"summary":"获取当前登录用户信息","tags":["系统账户管理"]}},"/api/accounts/{id}":{"get":{"operationId":"AccountsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"根据ID获取账户详情","tags":["系统账户管理"]},"patch":{"operationId":"AccountsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccountDto"}}}},"responses":{"200":{"description":""}},"summary":"更新账户信息","tags":["系统账户管理"]},"delete":{"operationId":"AccountsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"删除账户","tags":["系统账户管理"]}},"/api/roles":{"post":{"operationId":"RolesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["Roles"]},"get":{"operationId":"RolesController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Roles"]}},"/api/roles/{id}":{"get":{"operationId":"RolesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Roles"]},"patch":{"operationId":"RolesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["Roles"]},"delete":{"operationId":"RolesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Roles"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}}},"info":{"title":"接口文档","description":"接口文档","version":"v1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"accessToken":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateAccountDto":{"type":"object","properties":{"username":{"type":"string","description":"登录名","example":"admin"},"nickname":{"type":"string","description":"昵称","example":"管理员"},"password":{"type":"string","description":"密码","example":"Abc123456"},"telephone":{"type":"string","description":"手机号","example":"13800138000"},"email":{"type":"string","description":"邮箱","example":"admin.example.com"},"position":{"type":"string","description":"职位","example":"管理员"},"roleIds":{"description":"角色IDS","example":[1,2],"type":"array","items":{"type":"string"}}},"required":["username","password","roleIds"]},"UpdateAccountDto":{"type":"object","properties":{}},"CreateRoleDto":{"type":"object","properties":{}},"UpdateRoleDto":{"type":"object","properties":{}},"LoginDto":{"type":"object","properties":{}}}}}